rename `sign` subcommand to `signer`
