+++ tooling/cli.rs/src/signer.rs
