+++ tooling/cli.rs/src/signer/generate.rs
