+++ tooling/cli.rs/src/signer/sign.rs
